fix(core): drop leading slash from paths in withVCSGeneratedPaths API
This fix comes from accidentally add a leading slash in the generated
paths during fix the Elixir issue like this:

```go
modName := "potato" // assume the name is this.
dag.GeneratedCode(directory).
    WithVCSGeneratedPaths([]string{"/src/"+modName+"/lib/mix/tasks/dagger.invoke.ex"})
```

The `.gitattributes` produces as:

```
//src/potato/...
```

It's happens because the API prepending it without detecting the path
that user given. This fixes by dropping it if the path already
contained.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>