ci: Extract module runtime testing from test-modules
This splits test-modules further because it's the test suite with the
most failures since the last split. The intention is to keep doing this
until:
- all test suites complete within 10 mins (95%)
- all tests suites succeed 99% of the runs

Also, since test-modules is now much smaller, and test-everything-else
is larger, we need to split test-everything-else into two parts:
- test-engine & test-cli
- everything else

This is an intermediary state until we figure out a better arrangement.
The thing that we care about now is stability & time-to-completion.
Ideally, all of these should be a single test suite.

I also dropped "Dagger" from a bunch of test suites: it's all Dagger.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

Squash -2

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>