ci: publish engine variants from inside dagger
We move all of the publishing of the engine image into dagger - this
allows us to very easily implement functionality that first builds all
the variants, and only once this has succeeded, then can publish all the
variants. This helps prevent the case where some variants may not build,
and we end up in an invalid engine publishing state.

To do this reorganization, we need to re-organize some of the dev
module's engine api. Specifically, we need to lift the base image and
gpu-support parameters into *just* the `Container` method, instead of
parameterizing it across the entire engine object. This is so that we
can then easily use a go table of targets, and use this to parameterize
the calls to `Engine.Container`.

Signed-off-by: Justin Chadwell <me@jedevc.com>