ci: split `TestModule` into multiple suites (#8212)
* feat: split `TestModule` into multiple suites

Signed-off-by: Tom Chauveau <tom@epitech.eu>

* ci: Extract module runtime testing from test-modules

This splits test-modules further because it's the test suite with the
most failures since the last split. The intention is to keep doing this
until:
- all test suites complete within 10 mins (95%)
- all tests suites succeed 99% of the runs

Also, since test-modules is now much smaller, and test-everything-else
is larger, we need to split test-everything-else into two parts:
- test-engine & test-cli
- everything else

This is an intermediary state until we figure out a better arrangement.
The thing that we care about now is stability & time-to-completion.
Ideally, all of these should be a single test suite.

I also dropped "Dagger" from a bunch of test suites: it's all Dagger.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* chore: Fix merge conflict

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Split testdev into smaller suites

- testdev-modules
- testdev-module-runtimes
- testdev-container

So that tests are more likely to succeed. These instance sizes seem
excessive, should configure smaller ones as a follow-up.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Remove unused import

It was making the tests fail.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Tom Chauveau <tom@epitech.eu>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Tom Chauveau <tom@epitech.eu>
Co-authored-by: Gerhard Lazu <gerhard@dagger.io>