sdk(go): simplify DaggerObject aliasing slightly
This gets rid of the multiple levels of type aliasing used to be just
one level.

This is a very mild simplification that worst case seems harmless, but
the motivation right now is a shot in the dark attempt at seeing if it
avoids some of the flakes we have been getting in the Go SDK with this
object.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>