fix: don't skip collecting resource IDs in constructors
There was a conditional in the code to skip collecting resource IDs
(secrets, sockets, etc.) when the function had no name. I believe this
was originally meant to skip doing this for the call to get the module
typedefs.

However, it ended up also skipping that codepath for constructors, which
is wrong. Secrets and other resources can be made during constructor
calls that should be collected, otherwise you will get errors about
missing client resources if they are made in that call.

As far as I can tell, that check can just be removed entirely without
any harm, so that's the fix here.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>