fix: make Git.tags url parsing consistent with Git.tree
Previously, extracting the git tags from a URL skipped a level of url
pre-processing that was in buildkit (essentially, one that allowed no
protocols in URLs). We need to add this level of pre-processing
manually.

Without this, parsing module refs of the form "github.com/<org>/<repo>@<version>" would fail with:

	! failed to resolve git tags: select: git command failed: exit status 128
	Error: failed to get module ref kind: input: moduleSource resolve: failed to resolve git tags: select: git command failed: exit status 128
	stdout:
	stderr: fatal: 'github.com/<org>/<repo>' does not appear to be a git repository
	fatal: Could not read from remote repository.

This is because it's not a repo - it needs that same processing to be
consistent with `Git.Tree`.

Signed-off-by: Justin Chadwell <me@jedevc.com>