ci: fix schema reference generation
The schema reference generation takes the schema.graphqls file as input,
so each generate step effectively required two passes. This is kinda
messy, so we should just feed one generation step into the next.

This is kinda neat, since the dagger engine will deduplicate the schema
generation, so we don't end up computing it and generating it twice.

Signed-off-by: Justin Chadwell <me@jedevc.com>