💥 Breaking(helm): Remove appVersion - this is now the chart version (#8348)
* breaking(helm): Remove appVersion - this is now the chart version

Now that the chart version moves in lock-step with the app version, we
want to remove the app version so that we minimize the change for
incompatibilities, as was the case in
https://github.com/dagger/dagger/issues/8301.

The options are:
A. You want a custom app image. You need to ensure that the chart is
compatible with this image.
B. You use the default image. We are responsible for ensuring compatibility
with the chart.

The third option is to use this YAML as inspiration, and manage your own
raw manifests, which give you full control & responsibility.

Fixes https://github.com/dagger/dagger/issues/8301

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Make the different values stand out more and be easier to configure

Also added bunch of extra comments & fixed a link.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Lower the readiness timeout seconds to be less than the interval

If we are checking readiness every 15 seconds, we should timeout after
14 seconds at most.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Use dagger core version

Now that we have settled https://github.com/dagger/dagger/pull/8135,
let's use it.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: remove appVersion from Helm.setVersion

Signed-off-by: Justin Chadwell <me@jedevc.com>

* Clarify `engine.image.ref` setting & responsibility

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Bump the golangci-lint timeout to 10m

Our linting needs more than 5m.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>