ci: Move the Engine & CLI split test suites in a separate workflow (#8407)
So that we can compare the two different workflows side-by-side. This is
something that we have learned ~2 weeks after the last test suite split:
- https://github.com/dagger/dagger/pull/8212

We know that the split suites are faster & more reliable, but without
this change, we cannot compare the two workflows side-by-side. All that
we know is that the Engine & CLI workflow is failing more often (flakes
are now 2x as likely to happen). This change will show us how all the
tests, but split differently, compare.

This is part of a wider effort captured in:
- https://github.com/dagger/dagger/issues/8184

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>