add telemetry golden test suite
Finally, we have coverage for the full Engine -> OTel -> Frontend
pipeline, with examples and failure modes that are easy to understand.

Usage:

  dagger call test telemetry
  dagger call test telemetry --update -o . # update examples

TODO: why are the examples wrong?

TODO: did this _really_ need to be a separate suite + check? felt more
sane at the time, re-validate. originally the catch was --update but
maybe we just need a `golden` entrypoint? oh, but that would mean a
separate check anyway.

Signed-off-by: Alex Suraci <alex@dagger.io>