core: support GODEBUG as system env var + disable go mod index in CI
GODEBUG=goindex=0 is an undocumented feature flag that disabled Go's
internal module index. In an upstream issue we opened with Go around a
bug in build caching that's impacting us, they used it to workaround the
bug.

This change thus adds support for GODEBUG as a system env var (similar
to our support for GOPROXY) and uses that to set goindex=0 just in our
CI test suite for now.

This bug seems to only happen under extreme use cases such as our CI
where we're running many, many go module builds, which is why this
change is only being used in our CI so far (as opposed to setting
GODEBUG=goindex=0 for all production users by default).

If this ends up impacting real-world users and we get confirmation from
the Go devs that setting this feature flag doesn't cause any subtle
problems, we can consider setting it by default too. Hopefully though
the upstream bug is just fixed and we can undo all this.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>