tests: deflake TestPlatform/TestCacheMounts
This is all theoretical still, but the flakes we see in the test would
be explained by actual expected behavior: cache mounts are best effort
and free to be pruned at any time. If the cache mount used in this test
is pruned in the middle of it (between execs that use it), then we'd
expect to see the errors we're getting of `No such file or directory`.

This attempts to deflake the test by running an exec in the background
that has the cache volume mounted throughout the test.

Other possible mitigations were even more complicated (i.e. run a
separate engine with configuration that disabled cache volume pruning,
etc.), so while this feels a little weird I can't think of anything
better at this time other than removing the test entirely.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>