engine: log buildkit errors in right place to reveal stack traces
Previously we started logging buildkit solve errors with a `%+v` in
order to reveal the stack traces embedded in the errors by buildkit's
usage of github.com/pkg/errors, which in many places in buildkit is the
only way to get any extra information on where an error is actually
coming from (needed for debugging "context cancelled" errors
specifically).

This did not quite work for a pretty dumb reason: it turns out that if
you wrap an error from github.com/pkg/errors using fmt.Errorf (i.e. the
standard library) then the trick for extracting the stack trace from
the github.com/pkg/errors annotated error doesn't work anymore.

This is all quite silly, but in the interest of being able to debug
these context cancelled errors the change in this commit moves the log
statements of the errors with `%+v` to places *before* we do any
wrapping with the standard library, allowing us to see the stack traces.

I tested this by modifying buildkit locally to return errors from the
compute cache key step where we see this flake in CI and then verifying
the logs actually reveal the stack trace now.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>