engine: extract buildkit error stack trace lines into err msgs
There's reluctance upstream in buildkit to include messages in each
context cancel cause error, so instead we have to extract out the
embedded stack trace in each error, attempt to parse a line number from
the stack trace and include that in the user facing error message.

This keeps the user facing error somewhat succinct (i.e. not a huge
stack trace) while still making it a bit easier to debug this if users
hit this and can't/won't send us their full engine logs.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>