reveal linked spans, but mark execs passthrough
we don't always want to hide linked spans: Dockerfiles for example are
an example of a one-to-many cause => new effects relationship.

withExec -> exec is one-to-one, so in that case we want to hide it, but
we can just do that by setting Passthrough.

at this point we should probably stop using hacky [internal] prefixes
for names and just set attributes like we've done here.

Signed-off-by: Alex Suraci <alex@dagger.io>