ci: fix version module when no git metadata is available
If an empty directory was passed to the version module, then `.git`
wouldn't be available, and so all the git operations would fail. The fix
is to detect it before running any of the commands.

But to do that, we need a context and an error - so we just move all
this operation straight into the constructor, which also avoids a ton of
extra calls to the `Git` function anyways.

Signed-off-by: Justin Chadwell <me@jedevc.com>