chore: upgrade go to 1.23.2
I was getting ephemeral panics from the Go runtime on my OTEL metrics PR
that (seemingly?) disappeared when I upgraded to the new version of Go.
Not sure precisely which change fixed it (they were "invalid pointer
found in heap" errors) but there were a few fixes to the runtime and
open issues that seemed possibly related, so worth an upgrade.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>