refactor: module execution system to true imports
Rework the way the user's code is imported to be executed to actually
uses `imports` instead of loading it from its decorators.

This would unblock severals features such as interface supports and
compositions.

It's also a safer way to actually executes the user's code and resolve
default value by reference, specially complex values since
they can now be resolve by actually executing them if they are exported.

/!\ This is a breaking change! Anything that needs to be executed by
the runtime or the introspector now needs to be exported.

Signed-off-by: Tom Chauveau <tom@epitech.eu>