core: support disabling automatic init in withExec
Dagger automatically injects an init process in withExecs in order to
avoid surprising behavior around processes being PID 1. I have generally
gathered that this is the right default behavior, to the point that I've
seen some docker maintainers say that they wish docker had defaulted to
this same behavior.

However, there are less common use cases that do call for user processes
being PID 1, such as when they want very custom init behavior or are
testing init systems (i.e. testing systemd in a container).

For a long time this was impossible because of the shim but after we
moved to the executor, which operates entirely in the engine process, it
became technically optional to have an init process injected into the
container (even if the right default).

This change adds an option to `withExec` for disabling the automatic
init process, `noInit` so we can support those use cases. The default
behavior is left alone.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>