docs: sanitize comments for typescript doctype
Recently we added "${VAR}" snippets into our docs - however, as doctype
was generating MDX output, this was weirdly being interpreted as
variable interpolations.

There is a setting to control this: `sanitizeComments`. However, this
is only present in newer versions of the doctype plugin, so, a whole
upgrade of doctype and its docusaurus plugins is required, so I've done
that here too.

Signed-off-by: Justin Chadwell <me@jedevc.com>