chore(go): deprecated unscoped enum values
In go, enum values are top-level to the package (unlike in
typescript/python/etc where they are members of the type itself).
Because of this, it's very easily possible to clash - two different
enum types might have the same value.

To prevent this clash, we should scope the enum value with the name of
the enum it's from - this is pretty standard go practice, and should be
familiar.

We keep the old consts around, but alias them to the new consts, and
deprecate them (to remove in v0.14.0).

Signed-off-by: Justin Chadwell <me@jedevc.com>