ci: implement Alpine module w/ go apk lib for better caching
The previous module implementation called apko in a withExec, which
resulted in every permutation of a given set of alpine packages creating
a new tarball on disk (which added up to multiple GBs in our CI) and
importing those tarballs to their own, single-layer containers.

The new implementation uses the apk go bindings inside apko to directly
download, unpack and merge packages together all natively in Dagger.

The end result is that we only download+unpack packages once and use
merge-op to efficiently combine them into different permutations of
package installs, getting rid of the previous duplication and improving
caching.

This just updates our in-repo Alpine module for now. We should also do
the same for the Wolfi module so we can get rid of all usages of
wrapping apko in a withExec.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>