engine: cache internal sdk constructor calls
We were missing a check for IsInternal on module constructor calls
specifically (we had them on other function calls). This meant that we
never cached calls to SDK module constructors (i.e. any non-Go SDK).

For our CI specifically, this added ~1s of overhead to the `initialize`
step even when everything is cached.

Obviously 1s is not dramatic, but if we stack enough improvements like
this up we should see meaningful improvements in initialize time.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>