engine: compress SDK images w/ zstd to reduce image size + init time
Previously the SDK images included with the engine image were left
uncompressed. The idea was that the "outer" compression of the engine
image would handle compression and thus no extra layer was needed on the
cached SDK images.

However, it turns out that compressing the SDK images w/ Zstd reduces
the final engine image size and, more surprisingly, reduces uncached
module initialization time.

The engine image size is likely reduced just because the "outer"
compression is gzip but zstd is just better (simultaneously faster and
better compressed).
* `engine.tar` goes from 197MB to 191MB
* It would be nice to compress the whole engine image w/ zstd, but we've
  had problems with zstd compatibility for users on old versions of
  docker, so not pursuing that right now.

I was more surprised by the fact that the time to initialize modules in
the completely-uncached case went down too.
* Initializing our full CI module went from 1m13s to 1m7s
* Initializing `modules/alpine` by itself went from 23s to 21s

I haven't added enough spans to pinpoint why this is precisely (will do
in various follow-ups) but I'd guess that loading the SDK image from the
engine's filesystem into the content-store has enough overhead that the
smaller image actually helps even though we need to unpack it.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>