feat(sdk/dotnet): implements SDK pipeline
Add Dotnet to the `SDK` type and implements 3 functions, `lint`,
`test` and `generate`. Those functions are calling `dotnet-sdk-dev`
module that living inside `sdk/dotnet/dev`.

The custom dagger code in dev module and introspection.graphql are also
removed since we're rely on Dagger Engine on the root module.

And also add a dotnet pipeline to the GHA to make it run on CI.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>