feat(sdk/elixir): move entrypoint invoke task to dagger_sdk
Before this changeset, the invoke task living inside a user defined
module. This is confuses for someone who read the module source code.

So, this changeset move the invoke task to the `dagger_sdk` and modify
the runtime to use this task instead.

And the runtime is no longer generate mix task anymore.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>