engine: fix excessive cache invalidation for non-function nested execs
Before this, when nesting was enabled for an exec we would always scope
the cache key with the dagql call digest for that exec.

This is important for the case of a module function call since it
results in the cache key being scoped to that function's args and
receiver.

However, this was also happening for non-function nested execs (i.e. use
of the ExperimentalPrivilegedNesting flag). For those execs, this was
unnecessary. It also resulted in excessive cache invalidation since
dagql call digests aren't (yet) smart enough to incorporate
content-based cache keys.

Only enabling the cache key scoping for function calls specifically
fixes the problem.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>