engine: fix cache volume re-use when sourced from builtinContainer
I noticed that there were a bunch of engine logs about not being able to
re-use cache mount refs due to missing descriptors for lazy blobs.

These is buildkit mumbo jumbo that means it left certain cache refs lazy
but when it needed to unlazy them it didn't find the required metadata
to do so hidden in the context.

This was happening specifically because of the recent change to set go
module building containers to use cache volumes sourced from the base
container. The base container was loaded using OCI LLB which leaves
cache refs lazy.

There's an upstream bug involved here, but we can workaround in the
meantime by forcing an unlazy of builtin container cache refs.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>