ensure authToken secret is in dagql Call ID so it is passed to functions
In order for a function to be granted selective access to a secret, the
secret needs to appear in the dagql Call. Previously the secret was just
set on a field in the returned struct, now it is in the returned
dagql.Instance and thus passes to function clients through args.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>