tests: deflake cache volume tests from pruning
I saw a few of these tests flake today due to cache volumes not having
contents, which would be entirely possible by the cache volume being
pruned by buildkit in between usage. There was a recent fix to a
platform test to handle this and it hasn't flaked since; just applying
that fix to a few more tests now.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>