tests: disable go test verbosity by default
the -v flag on go test results in all output being printed, whereas the
default behavior only shows output when a test fails.

The majority of the time we only care about output from failed tests, so
we are usually wasting time processing + uploading tons of output from
tests.

This change disables -v by default but leaves it as an option to enable
when needed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>