go: delay old enum name removal until v0.15.0
There was very little time the old names were deprecated (like a week)
and because their removal is automagical we didn't notice it until the
last minute and thus didn't have time to prep any notes on it.

So we decided to delay removing the deprecated names until the next
patch release.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>