ci: Fix the daggerverse-preview trigger
TIL: `/` only applies to git (not GitHub)

The semantics are: if a new PR gets created which changes the
`.changes/.next` file, deploy a preview environment.

These PRs **must** be opened from branches that belog to the upstream
repository - `dagger/dagger` - otherwise the secret will not be
available and therefore the `dagger/dagger.io` PR cannot be created due
to insufficient permissions. I have updated `RELEASING.md` to reflect
this.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>