engine: wrap executor cleanup errors
While trying to debug a flake I was suspecting a cleanup step might be
failing. Problem is all I'm getting is "context deadline exceeded", so
if it is a cleanup step, I can't tell which one.

This updates any cleanup errors to be wrapped with the context on what
specifically failed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>