engine: account for oci index in TestFromImagePlatform
This test started failing out of band in our CI on amd64 hosts.

Justin noticed that the published arm64v8/alpine image had changed from
a single manifest to a multiplatform oci index w/ a single
platform-specific manifest.

This broke the test because previously in the "single manifest" case the
platform we tried to resolve the image with was completely ignored.
After switching to the oci index, the platform actually mattered.

We were defaulting the platform to the host's platform when resolving
the image, but this test purposely attempts to pull an image for a
different platform.

After some discussion we decided erroring in the case is expected and
the simplest behavior. If users want to pull images that don't match
their host's platform, they need to explicitly request that platform in
the Container opts.

This thus updates the test case to explicitly ask for the desired
platform.

I also included a small improvement to the error messages when hitting
this case, which would have sped up debugging this a bit.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>