go: wait for CLI session cmd to finish copying on close
I saw the race detector go off in a test run where it appeared to be
saying that the Go SDK CLI session code was still finishing an io.Copy
to the LogOutput for the test *after* the test was done.

That io.Copy was running in an unsynchronized goroutine, so this seems
entirely plausible in the case where it is still reading some buffered
data from the stderr pipe after the child process exited.

This updates the SDK client's Close method to wait for that io.Copy to
finish before returning, which should ensure there's no stray writes
after Close ends.

I couldn't repro this test race locally after many attempts, so this is
still a bit of an educated guess. Won't know for sure it's gone until we
haven't seen it in CI for a bit.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>