engine: make otel proxy shutdown non-fatal
I haven't tracked down why, but it appears that sometimes in our test
workflows the "shutdown otel proxy" cleanup step takes over 10s, hits
its timeout and then causes the exec to fail.

It shouldn't take that long, but in the meantime there's no reason for
this to be fatal to the exec. The error is now just logged.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>