engine: make otel proxy shutdown error non-fatal (#8936)
* engine: wrap executor cleanup errors

While trying to debug a flake I was suspecting a cleanup step might be
failing. Problem is all I'm getting is "context deadline exceeded", so
if it is a cleanup step, I can't tell which one.

This updates any cleanup errors to be wrapped with the context on what
specifically failed.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: make otel proxy shutdown non-fatal

I haven't tracked down why, but it appears that sometimes in our test
workflows the "shutdown otel proxy" cleanup step takes over 10s, hits
its timeout and then causes the exec to fail.

It shouldn't take that long, but in the meantime there's no reason for
this to be fatal to the exec. The error is now just logged.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>