engine: set synchronous=OFF for client DBs
All of this data is ephemeral - we don't care if we lose it on a system
crash. Setting synchronous=NORMAL opens up to slowdowns from random I/O
blocks, as observed when running `test all`.

Without this change we could occasionally see a span insert blocking for
more than 10 seconds, triggering a timeout when shutting down the OTel
proxy server. The odds of this happening in `test all` were pretty high
- at least one in two runs.

After this change I have been able to run `test all` 8 times and none of
them hit that particular issue. (Three failed, but that's neither here
nor there.)

Signed-off-by: Alex Suraci <alex@dagger.io>