tests: fix terminal tests not asserting as expected
Our terminal tests were not actually working how we thought they were.
Many of them were not waiting for prompts and instead just matching on
strings right away, which ended up including all the progress output
before the terminal was opened.

This broke a few of them since the string they were expecting happened
to also show up in the progress output, which isn't what we actually
wanted to assert on.

This updates the tests to correctly wait for prompts (doing some
terminal escape sequence handling when needed).

This commit leaves the `TestDaggerTerminal/directory` test case failing
because that feature was actually totally broken due to a separate issue
(which went unnoticed until now because the test case was broken).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>