engine: fix Directory.terminal and overlay mount handling
The previous commit fixed assertions in our terminal test cases, which
revealed that the Directory.terminal feature was totally broken.

The problem was in the executor and it's handling of mounts that were of
type overlay. It was looking at the `Source` field of the mount assuming
it was a bind and that the field pointed to the source of the bind.

However, for non-bind mounts the `Source` is just the type of mount.

The reason this only broke the Directory.terminal case so far was that
Directory.terminal is the only place in our code where we create
read-only mounts. This triggers a different codepath in buildkit's cache
logic that results in actual overlay mounts being provided to the
executor; in other read-write mount cases buildkit has some logic that
creates a "shared" overlay mount which is just bind-mounted to other
places (to avoid conflicts around multiple overlay mounts with the same
upperdir).

This commit just fixes handling of that `Source` field to interpret it
correctly based on whether it's a bind mount or not.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>