Prevent occasional OTel SQLite writing slowdowns (#8950)
* engine: set synchronous=OFF for client DBs

All of this data is ephemeral - we don't care if we lose it on a system
crash. Setting synchronous=NORMAL opens up to slowdowns from random I/O
blocks, as observed when running `test all`.

Without this change we could occasionally see a span insert blocking for
more than 10 seconds, triggering a timeout when shutting down the OTel
proxy server. The odds of this happening in `test all` were pretty high
- at least one in two runs.

After this change I have been able to run `test all` 8 times and none of
them hit that particular issue. (Three failed, but that's neither here
nor there.)

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert "don't bother with transactions for client DB (#8861)"

This reverts commit 925cf02d75b452fb0be2bd77f0da10cbd1872105.

This didn't end up helping the one user it affected, and it feels better
for each request to be all-or-nothing considering OTel will retry
requests.

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>